refactor policy output and cliff impact
